+++ RELEASES.md
@@ -7,3 +7,11 @@ Based on:
 - Speakeasy CLI 1.520.1 (2.556.1) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [typescript v2.0.0-alpha.0] .
+
+## 2025-03-27 00:15:06
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.524.0 (2.562.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.0.0-alpha.1] .